Add delete path if tour stop is deleted

+++ app/controllers/tour_stops_controller.rb
@@ -15,8 +15,8 @@ class TourStopsController < ApplicationController
   def destroy
     @tour_stop = TourStop.find params[:id]
     path = @tour_stop.stop_type.classify.constantize.find_by_id(@tour_stop.stop_id).paths.last
-    path.path_points.destroy_all
-    path.destroy
+    path.path_points.destroy_all if path.present?
+    path.destroy if path.present?
     VisitedStop.where(tour_stop_id: @tour_stop.id).destroy_all
     if @tour_stop.destroy
       redirect_to community_tours_path(current_community), :notice => "Tour Stop deleted"
